Final Version(?)
I think this has everything in it that we wanted. The program should now prevent the user from choosing the same arrival city as destination city (this part was weird and I'm not really sure why it works but it does lol). Also, the program checks if the user's arrival date is before the departure date (also hard but I understand this one). Hopefully it works the same on linux as it does on windows, otherwise it may need to be revised. If someone can test, that would be awesome.